requirements.txt
PyQt5
launchdarkly-server-sdk
requests
pyyaml
